requirements.txt
matplotlib==3.10.8
numpy==2.4.2
optuna==4.7.0
pandas==3.0.1
pandas_ta==0.4.71b0
psycopg2_binary==2.9.11
python-dotenv==1.2.2
python_dateutil==2.9.0.post0
